Rename TagsHelpers.ts to frames-reader.ts
The content of this module is now only dedicated to reading frames.
